Move some files to `sharedTest` and reformat imports
