+++ .codeclimate.yml
@@ -1,4 +1,5 @@
 exclude_patterns:
+  - "**/sharedTest/"
   - "**/test/"
   - "**/androidTest/"
   - "**test"
