+++ mobile/src/androidTest/java/ch/epfl/sdp/mobile/androidTest/state/NavigationTest.kt
@@ -6,18 +6,18 @@ import androidx.compose.ui.test.*
 import androidx.compose.ui.test.SemanticsMatcher.Companion.expectValue
 import androidx.compose.ui.test.SemanticsMatcher.Companion.keyIsDefined
 import androidx.compose.ui.test.junit4.createComposeRule
-import ch.epfl.sdp.mobile.application.ProfileDocument
-import ch.epfl.sdp.mobile.application.authentication.AuthenticationFacade
-import ch.epfl.sdp.mobile.application.chess.ChessFacade
-import ch.epfl.sdp.mobile.application.social.SocialFacade
-import ch.epfl.sdp.mobile.state.Navigation
-import ch.epfl.sdp.mobile.state.ProvideFacades
 import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.auth.SuspendingAuth
 import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.auth.buildAuth
 import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.auth.emptyAuth
 import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.store.buildStore
 import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.store.document
 import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.store.emptyStore
+import ch.epfl.sdp.mobile.application.ProfileDocument
+import ch.epfl.sdp.mobile.application.authentication.AuthenticationFacade
+import ch.epfl.sdp.mobile.application.chess.ChessFacade
+import ch.epfl.sdp.mobile.application.social.SocialFacade
+import ch.epfl.sdp.mobile.state.Navigation
+import ch.epfl.sdp.mobile.state.ProvideFacades
 import kotlinx.coroutines.test.runTest
 import org.junit.Rule
 import org.junit.Test
