+++ mobile/src/androidTest/java/ch/epfl/sdp/mobile/androidTest/ui/authentication/AuthenticationScreenTest.kt
@@ -5,15 +5,15 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createComposeRule
+import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.auth.buildAuth
+import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.auth.emptyAuth
+import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.store.emptyStore
+import ch.epfl.sdp.mobile.androidTest.state.setContentWithLocalizedStrings
 import ch.epfl.sdp.mobile.application.authentication.AuthenticationFacade
 import ch.epfl.sdp.mobile.application.chess.ChessFacade
 import ch.epfl.sdp.mobile.application.social.SocialFacade
 import ch.epfl.sdp.mobile.state.ProvideFacades
 import ch.epfl.sdp.mobile.state.StatefulAuthenticationScreen
-import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.auth.buildAuth
-import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.auth.emptyAuth
-import ch.epfl.sdp.mobile.androidTest.infrastructure.persistence.store.emptyStore
-import ch.epfl.sdp.mobile.androidTest.state.setContentWithLocalizedStrings
 import ch.epfl.sdp.mobile.ui.authentication.AuthenticationScreen
 import ch.epfl.sdp.mobile.ui.authentication.AuthenticationScreenState
 import ch.epfl.sdp.mobile.ui.authentication.AuthenticationScreenState.Mode.*
